# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CISync Control-Plane API
  version: 1.0.0
  description: >
    Agent-native verification & integration control plane. Normative contract for v1.
    All mutating requests require an Idempotency-Key header; replays return the original
    response (I-12). Tenant identity derives from the Authorization token, never payload.
servers:
  - url: http://localhost:8081/v1
security:
  - bearerAuth: []
paths:
...
  /change-intents/{intentId}:
    get:
      operationId: getIntent
      parameters: [{ $ref: '#/components/parameters/IntentId' }]
      responses:
        '200':
          description: Intent state incl. evidence_completeness_pct (D8 formula)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Intent' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /candidates/{candidateId}:
    get:
      operationId: getCandidate
      parameters: [{ $ref: '#/components/parameters/CandidateId' }]
      responses:
        '200':
          description: Candidate + priority + queue_position
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Candidate' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /events:
    get:
      operationId: tailEvents
      parameters:
        - { name: after_seq, in: query, required: true, schema: { type: integer, format: int64, minimum: 0 } }
        - { name: types, in: query, schema: { type: string, description: comma-separated event types } }
        - { name: aggregate, in: query, schema: { type: string, description: '"type:id" filter' } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
      responses:
        '200':
          description: Ledger tail for agents/web sync
          content:
            application/json:
              schema:
                type: object
                required: [events, next_seq]
                properties:
                  events:
                    type: array
                    items: { $ref: '#/components/schemas/EventEnvelope' }
                  next_seq: { type: integer, format: int64 }
        '416': { $ref: '#/components/responses/BadRequest' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    IntentId:
      name: intentId
      in: path
      required: true
      schema: { type: string, pattern: '^int_[0-9A-HJKMNP-TV-Z]{26}$' }
    CandidateId:
      name: candidateId
      in: path
      required: true
      schema: { type: string, pattern: '^cand_[0-9A-HJKMNP-TV-Z]{26}$' }
...
  responses:
    BadRequest:
      description: validation_failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Uniform 404 (never distinguishes cross-tenant existence)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
    RiskClass: { type: string, enum: [low, medium, high, critical] }
...
    IntentState:
      type: string
      enum: [exploring, validating, blocked, repairing, merge_ready, deploying, monitoring, completed, rejected]
    Intent:
      type: object
      required: [intent_id, state, goal, repository, owned_surfaces, risk_class, evidence_completeness_pct, created_at]
      properties:
        intent_id: { type: string }
        state: { $ref: '#/components/schemas/IntentState' }
        goal: { type: string }
        repository: { type: string }
        owned_surfaces: { type: array, items: { type: string } }
        risk_class: { $ref: '#/components/schemas/RiskClass' }
        evidence_completeness_pct: { type: number, minimum: 0, maximum: 1 }
        deadline: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        closed_at: { type: string, format: date-time, nullable: true }
    CandidateSummary:
      type: object
      required: [candidate_id, state, head_sha, cluster_id, relation_to_rep]
      properties:
        candidate_id: { type: string }
        state: { type: string }
        head_sha: { type: string }
        priority_score: { type: number }
        cluster_id: { type: string, nullable: true }
        relation_to_rep:
          type: [string, 'null']
          enum: [duplicate_of, alternative_of, composable_with, conflicts_with, prerequisite_of, null]
...
    Candidate:
      allOf:
        - $ref: '#/components/schemas/CandidateSummary'
        - type: object
          required: [intent_id, queue_position, est_cost_millicents]
          properties:
            intent_id: { type: string }
            queue_position: { type: integer, nullable: true }
            est_cost_millicents: { type: integer }
...
    EventEnvelope:
      type: object
      required: [id, seq, type, version, tenant_id, aggregate, causation_id, correlation_id, actor, occurred_at, payload_sha256, prev_hash, entry_hash, payload]
      properties:
        id: { type: string, pattern: '^evt_[0-9A-HJKMNP-TV-Z]{26}$' }
        seq: { type: integer, format: int64 }
        type: { type: string, example: candidate.submitted }
        version: { type: integer }
        tenant_id: { type: string }
        aggregate:
          type: object
          properties:
            type: { type: string }
            id: { type: string }
        causation_id: { type: string }
        correlation_id: { type: string }
        actor:
          type: object
          properties:
            kind: { type: string, enum: [agent, human, service, github] }
            id: { type: string }
        occurred_at: { type: string, format: date-time }
        payload_sha256: { type: string }
        prev_hash: { type: string }
        entry_hash: { type: string }
        payload: { type: object }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [validation_failed, unauthorized, forbidden, not_found, conflict_state, rate_limited, budget_exceeded, unavailable]
            message: { type: string }
            details: { type: object, additionalProperties: true }
            retry_after_s: { type: integer, nullable: true }
            suggestions: { type: array, items: { type: string } }
...